% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacement.R
\name{find_displacement}
\alias{find_displacement}
\title{displacement}
\usage{
find_displacement(v, t, a)
}
\arguments{
\item{v}{initial velocity}

\item{t}{time}

\item{a}{acceleration}
}
\value{
finds total displacement
}
\description{
displacement
}
