% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_displacement.R
\name{plot_displacement}
\alias{plot_displacement}
\title{Plot_displacement}
\usage{
plot_displacement(v, t, a)
}
\arguments{
\item{v}{initial velocity}

\item{t}{time}

\item{a}{acceleration}
}
\value{
returns a plot of the distance traveled over time
}
\description{
Plot_displacement
}
