% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Velocity Squared.R
\name{V2}
\alias{V2}
\title{find_v_squared}
\usage{
V2(vi, a, d)
}
\arguments{
\item{vi}{Initial velocity}

\item{a}{acceleration}

\item{d}{displacement}
}
\value{
Velocity squared from the defined parameters
}
\description{
find_v_squared
}
