% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_velocity.R
\name{velocity}
\alias{velocity}
\title{find_velocity}
\usage{
velocity(v, a, t)
}
\arguments{
\item{v}{initial velocity}

\item{a}{acceleration}

\item{t}{time}
}
\value{
returns velocity at defined parameters
}
\description{
find_velocity
}
